fix-back : refactoring ; sonarqube 정적 분석을 토대로 npe가 발생할 수 있는 부분 위주로 수정합니다.

See merge request s08-webmobile2-sub2/S08P12A807!337

+++ togetdog/src/main/java/com/ssafy/togetdog/user/model/service/JwtService.java
@@ -5,8 +5,8 @@ import java.util.Map;
 public interface JwtService {
 	<T> String createAccessToken(long userId);
 	<T> String createAccessToken(long userId, String role);
-	<T> String create(long userId, String subject, long expir);
-	<T> String create(long userId, String subject, long expir, String role);
+	<T> String create(long userId, String subject, int expir);
+	<T> String create(long userId, String subject, int expir, String role);
 	void validateToken(String jwtToken);
 	long getUserId(String jwtToken);
 	Map<String, Object> get(String key);
